pyproject.toml
[tool.poetry]
name = "repl_python3_OutDashRewrite"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
quart = "^0.16.1"
hypercorn = "^0.12.0"
praw = "^7.5.0"
pymongo = "^4.0.1"
python-dotenv = "^0.19.2"
disnake = "^2.3.0"
dnspython = "^2.1.0"
aiohttp = "<3.8.0"
cachetools = "^4.2.4"
PyJWT = "^2.3.0"
async_oauthlib = "^0.0.9"
oauthlib = "^3.1.1"
asgiref = "^3.4.1"
flask-discord = "^0.1.64"
flask = "^2.0.1"
waitress = "^2.0.0"
api = "^0.0.7"
nest-asyncio = "^1.5.4"
httpx = "^0.21.1"
apiclient = "^1.0.4"
"randomstuff.py" = "^2.0.0"
youtube-dl = "^2021.12.17"
pynacl = "^1.4.0"
discord = "^1.7.3"
discord-webhook = "^0.14.0"
google-api-python-client = "^2.34.0"
captcha = "^0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"